@@ -1,7 +1,7 @@
 {
   "name": "mui-breakpoint-indicator",
   "version": "0.2.1",
-  "description": "shows you your current material-ui breakpoint. tiny. allows default position setting.",
+  "description": "shows you your current mui breakpoint. tiny. allows default position setting.",
   "main": "lib/index.js",
   "author": "Dimitri Mitropoulos <dimitrimitropoulos@gmail.com>",
   "license": "MIT",
@@ -12,15 +12,15 @@
     "build": "yarn tsc --declaration --skipLibCheck"
   },
   "dependencies": {
-    "@material-ui/core": "^4.11.0",
-    "polished": "^4.0.3",
-    "ramda": "^0.27.1",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "typescript": "^4.0.3"
+    "@mui/material": "^5.2.8",
+    "polished": "^4.1.3",
+    "ramda": "^0.27.2",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "typescript": "^4.5.4"
   },
   "devDependencies": {
-    "@types/ramda": "^0.27.28",
-    "@types/react": "^16.9.52"
+    "@types/ramda": "^0.27.62",
+    "@types/react": "^17.0.38"
   }
 }
