@@ -15,7 +15,7 @@ exports.BreakpointIndicator = void 0;
 const React = require("react");
 const react_1 = require("react");
 const polished_1 = require("polished");
-const core_1 = require("@material-ui/core");
+const material_1 = require("@mui/material");
 const ramda_1 = require("ramda");
 /**
  * taken from https://material-ui.com/components/use-media-query/#migrating-from-withwidth
@@ -26,12 +26,12 @@ const ramda_1 = require("ramda");
  */
 const useWidth = () => {
     var _a;
-    const theme = core_1.useTheme();
-    return ((_a = ramda_1.reduce((accumulator, breakpoint) => {
+    const theme = (0, material_1.useTheme)();
+    return ((_a = (0, ramda_1.reduce)((accumulator, breakpoint) => {
         // eslint-disable-next-line react-hooks/rules-of-hooks -- see above
-        const matches = core_1.useMediaQuery(theme.breakpoints.up(breakpoint));
+        const matches = (0, material_1.useMediaQuery)(theme.breakpoints.up(breakpoint));
         return accumulator === null && matches ? breakpoint : accumulator;
-    }, null, ramda_1.reverse(theme.breakpoints.keys))) !== null && _a !== void 0 ? _a : 'xs');
+    }, null, (0, ramda_1.reverse)(theme.breakpoints.keys))) !== null && _a !== void 0 ? _a : 'xs');
 };
 const getHue = (breakpoint) => {
     switch (breakpoint) {
@@ -71,11 +71,11 @@ const placementsByPosition = {
 };
 const InternalBreakpointIndicator = ({ className = '', position: suppliedPosition, style = {}, }) => {
     const defaultPosition = suppliedPosition !== null && suppliedPosition !== void 0 ? suppliedPosition : positionOrder[0];
-    const [position, setPosition] = react_1.useState(defaultPosition);
+    const [position, setPosition] = (0, react_1.useState)(defaultPosition);
     const breakpoint = useWidth();
     const hue = getHue(breakpoint);
-    const backgroundColor = polished_1.hsl(hue, 0.75, 0.5);
-    const color = polished_1.readableColor(backgroundColor);
+    const backgroundColor = (0, polished_1.hsl)(hue, 0.75, 0.5);
+    const color = (0, polished_1.readableColor)(backgroundColor);
     const onClick = event => {
         event.stopPropagation();
         const index = positionOrder.indexOf(position);
@@ -99,8 +99,9 @@ const InternalBreakpointIndicator = ({ className = '', position: suppliedPositio
     };
     return (React.createElement("div", { className: className, onClick: onClick, style: Object.assign(Object.assign({ backgroundColor, border: `2px solid ${color}`, color, cursor: 'pointer', padding: '0 1em', position: 'fixed', zIndex: 99999 }, placementsByPosition[position]), style) }, breakpoint));
 };
-exports.BreakpointIndicator = (_a) => {
+const BreakpointIndicator = (_a) => {
     var { visible = true } = _a, rest = __rest(_a, ["visible"]);
     return (visible ? React.createElement(InternalBreakpointIndicator, Object.assign({}, rest)) : null);
 };
+exports.BreakpointIndicator = BreakpointIndicator;
 //# sourceMappingURL=BreakpointIndicator.js.map
